remove placeholder files with empty test files
