+++ tests/unit/controller/utils/test_desktop_utils.py
@@ -1,0 +1,1 @@
+
